+++ scripts/macros/2024/spells/iceStorm.js
@@ -1,0 +1,6 @@
+import {iceStorm as iceStormLegacy} from '../../../legacyMacros.js';
+export let iceStorm = {
+    ...iceStormLegacy,
+    version: '1.2.29',
+    rules: 'modern'
+};